Move `ElectronicDelivery` components
