+++ .babelrcExp
@@ -55,8 +55,8 @@
           "@SearchButton": "./src/app/components/Buttons/SearchButton.jsx",
           "@SearchIconReversed": "./src/app/components/Buttons/SearchIconReversed.jsx",
           "@DataLoader": "./src/app/components/DataLoader/DataLoader.jsx",
-          "@ElectronicDelivery": "./src/app/components/ElectronicDelivery/ElectronicDelivery.jsx",
-          "@ElectronicDeliveryForm": "./src/app/components/ElectronicDelivery/ElectronicDeliveryForm.jsx",
+          "@ElectronicDelivery": "./src/app/pages/ElectronicDelivery.jsx",
+          "@ElectronicDeliveryForm": "./src/app/components/ElectronicDeliveryForm/ElectronicDeliveryForm.jsx",
           "@Feedback": "./src/app/components/Feedback/Feedback.jsx",
           "@FilterPopup": "./src/app/components/FilterPopup/FilterPopup.jsx",
           "@FieldsetDate": "./src/app/components/Filters/FieldsetDate.jsx",
